Add files via upload
OVERVIEW

The Online Examination System is a web application developed using Spring Boot, Thymeleaf, MySQL, and REST APIs. It allows users to create, manage, and take online exams, with functionalities for user registration, login, exam creation, and result viewing. This project is designed to provide a robust and scalable platform for conducting online exams with ease.